Consolidate blocked-user and start-chat screens into UserProfileScreen
