+++ lib/screens/user_search_screen.dart
@@ -106,7 +106,7 @@ class UserSearchScreen extends HookConsumerWidget {
                                 pictureUrl: user.metadata.picture,
                                 avatarColor: AvatarColor.fromPubkey(user.pubkey),
                                 size: WnUserItemSize.big,
-                                onTap: () => Routes.pushToStartChat(
+                                onTap: () => Routes.pushToUserProfile(
                                   context,
                                   user.pubkey,
                                 ),
